requirements.txt
mysql-connector-python
selenium
selenium-wire
requests
fastapi 
uvicorn 
python-dotenv 